add validtion layer to controllers for input

+++ src/features/task/task.controller.ts
@@ -1,12 +1,18 @@
 import { Response, Request } from "express";
 import * as taskService from './task.service'
-import { IUpdateData } from "../helpers/update.interface";
+import { IUpdateData } from "../utils/update.interface";
 import ITask from "./task.interface";
 import { Types } from "mongoose";
+import { handelValidation } from "../utils/common.validators";
 
 
 async function getTasksController(req: Request, res: Response): Promise<void> {
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const tasks = await taskService.getTasksService()
         res.json(tasks)
     }catch(err: any){
@@ -16,8 +22,14 @@ async function getTasksController(req: Request, res: Response): Promise<void> {
 
 async function getTaskController(req: Request, res: Response): Promise<void> {
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const id = req.params.task_id
         const task = await taskService.getTaskService(new Types.ObjectId(id));
+        if (!task) throw new Error('Task not found');
         res.json(task)
     }catch(err: any){
         res.status(404).json(err.message);
@@ -26,6 +38,11 @@ async function getTaskController(req: Request, res: Response): Promise<void> {
 
 async function createTaskController(req: Request, res: Response): Promise<void> {
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const reqBody = req.body;
         const taskData: ITask = {
             title: reqBody.title,
@@ -46,6 +63,11 @@ async function createTaskController(req: Request, res: Response): Promise<void> 
 
 async function updateTaskController(req: Request, res: Response): Promise<void> {
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const reqBody = req.body;
         const {task_id} = req.params;
         const data: IUpdateData = {
@@ -58,7 +80,7 @@ async function updateTaskController(req: Request, res: Response): Promise<void> 
             content: reqBody.content,
         }
         const task = await taskService.updateTaskService(new Types.ObjectId(task_id), data);
-        res.sendStatus(200);
+        res.sendStatus(204);
     }catch(err: any){
         res.status(400).json(err.message);
     }
@@ -67,6 +89,11 @@ async function updateTaskController(req: Request, res: Response): Promise<void> 
 async function deleteTaskController(req: Request, res: Response): Promise<void> {
     const taskId = new Types.ObjectId(req.params.task_id);
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const val = await taskService.deleteTaskService(taskId);
         if (val)
             res.sendStatus(204);
@@ -77,20 +104,30 @@ async function deleteTaskController(req: Request, res: Response): Promise<void> 
     }
 }
 
-function addFavoredUserController(req: Request, res: Response): void {
+async function addFavoredUserController(req: Request, res: Response): Promise<void> {
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const {task_id} = req.params;
         const {userId} = req.body;
-        taskService.addFavoredUserService(new Types.ObjectId(task_id), new Types.ObjectId(userId));
+        await taskService.addFavoredUserService(new Types.ObjectId(task_id), new Types.ObjectId(userId));
         res.sendStatus(204);
     }catch(err: any){
         res.status(400).json(err.message);
     }
 }
-function removeFavoredUserController(req: Request, res: Response): void{
+async function removeFavoredUserController(req: Request, res: Response): Promise<void>{
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const {task_id, user_id} = req.params;
-        taskService.removeFavoredUserService(new Types.ObjectId(task_id), new Types.ObjectId(user_id));
+        await taskService.removeFavoredUserService(new Types.ObjectId(task_id), new Types.ObjectId(user_id));
         res.sendStatus(204);
     }catch(err: any){
         res.status(400).json(err.message);
