+++ src/features/user/user.controller.ts
@@ -2,7 +2,8 @@ import { Response, Request } from "express";
 import * as userService from './user.service'
 import IUser from "./user.interface";
 import { Types } from "mongoose";
-import { IUpdateData } from "../helpers/update.interface";
+import { IUpdateData } from "../utils/update.interface";
+import { handelValidation} from "../utils/common.validators";
 async function getUsersController(req: Request, res: Response): Promise<void> {
     try{
         const users = await userService.getUsersService()
@@ -15,8 +16,14 @@ async function getUsersController(req: Request, res: Response): Promise<void> {
 
 async function getUserController(req: Request, res: Response): Promise<void>{
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const id = req.params.user_id
         const user = await userService.getUserService(new Types.ObjectId(id))
+        if (!user) throw new Error('User not found');
         res.json(user)
     }catch(err: any){
         res.status(404).json(err.message);
@@ -24,8 +31,13 @@ async function getUserController(req: Request, res: Response): Promise<void>{
 }
 
 async function createUserController(req: Request, res: Response): Promise<void>{
-    try{
+    try{  
         const reqBody = req.body;
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const userData: IUser = {
             username: reqBody.username,
             email: reqBody.email,
@@ -42,6 +54,11 @@ async function createUserController(req: Request, res: Response): Promise<void>{
 
 async function updateUserController(req: Request, res: Response): Promise<void>{
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const reqBody = req.body;
         const {user_id} = req.params;
         const data: IUpdateData = {
@@ -51,7 +68,7 @@ async function updateUserController(req: Request, res: Response): Promise<void>{
             last_name: reqBody.last_name,
         }
         const user = await userService.updateUserService(new Types.ObjectId(user_id), data);
-        res.sendStatus(200);
+        res.sendStatus(204);
     }catch(err: any){
         res.status(400).json(err.message);
     }
@@ -59,6 +76,11 @@ async function updateUserController(req: Request, res: Response): Promise<void>{
 
 async function deleteUserController(req: Request, res: Response): Promise<void>{
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const isDeleted = await userService.deleteUserService(new Types.ObjectId(req.params.user_id));
         if (isDeleted){
             res.sendStatus(204);
@@ -69,11 +91,11 @@ async function deleteUserController(req: Request, res: Response): Promise<void>{
         res.status(400).json(err.message);
     }
 }
-
 export {
     getUserController,
     getUsersController,
     updateUserController,
     deleteUserController,
-    createUserController
+    createUserController,
 }
+
